Conflicts:
	application/bootstrap.php
	application/classes/Ushahidi/Console/PostExporter.php
	application/classes/Ushahidi/Core.php
	application/classes/Ushahidi/Listener/ContactListener.php
	application/classes/Ushahidi/Repository/Dataprovider.php
	application/classes/Ushahidi/Repository/Form/Attribute.php
	application/classes/Ushahidi/Repository/Post/Export.php
	application/classes/Ushahidi/Validator/Tag/Update.php
	plugins/frontlinesms/classes/Controller/Sms/Frontlinesms.php
	src/Core/Entity/FormAttributeRepository.php
	src/Core/Entity/Permission.php
	src/Core/Exception/ValidatorException.php
	src/Core/Tool/Authorizer/SetAuthorizer.php
	src/Core/Tool/Uploader.php
	src/Core/Usecase/Media/CreateMedia.php
	src/Core/Usecase/User/ResetUserPassword.php
	tests/datasets/ushahidi/Base.yml

+++ src/Core/Usecase/Media/CreateMedia.php
@@ -14,7 +14,7 @@ namespace Ushahidi\Core\Usecase\Media;
 use Ushahidi\Core\Tool\Uploader;
 use Ushahidi\Core\Tool\UploadData;
 use Ushahidi\Core\Usecase\CreateUsecase;
-use Ushahidi\Exception\ValidatorException;
+use Ushahidi\Core\Exception\ValidatorException;
 
 class CreateMedia extends CreateUsecase
 {
@@ -46,7 +46,7 @@ class CreateMedia extends CreateUsecase
         } catch (ValidatorException $e) {
             // If a file was uploaded, it must be purged after a failed upload.
             // Otherwise storage will be filled with junk files.
-            if ($this->upload->file) {
+            if ($this->upload && $this->upload->file) {
                 $this->uploader->delete($this->upload->file);
             }
 
@@ -63,9 +63,15 @@ class CreateMedia extends CreateUsecase
     protected function getEntity()
     {
         // Upload the file and get the file reference
-        $this->upload = $this->uploader->upload(
-            new UploadData($this->getPayload('file'))
-        );
+        try {
+            $this->upload = $this->uploader->upload(
+                new UploadData($this->getPayload('file'))
+            );
+        } catch (\InvalidArgumentException $e) {
+            throw new ValidatorException($e->getMessage(), [
+                'file' => $e->getMessage()
+            ], $e);
+        }
 
         $payload = [
             'caption'    => $this->getPayload('caption', false) ?: null,
